shadow-claw/gateway/requirements.txt
python-telegram-bot==21.0.1
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
yfinance==0.2.40
